:sparkles: Adds virtual grip site link to launch file
In order to control the robot w.r.t. the center of the gripper a virtual grip site link and joint were added to the franka_ros/robot_description urdfs and the panda_moveit_config/config srdfs. In this commit additional arguments grip_site are added to the simulation launch file.